[IMP] Moved up export_case_json in the business layer
